Changed wheel amount type to int.

+++ client/Android/RemoteControl/app/src/main/java/net/mindwalkers/myapplication/MouseWheel.java
@@ -7,14 +7,13 @@ public class MouseWheel {
 
     @SerializedName("amount")
     @Expose
-    private Double amount;
+    private Integer amount;
 
-    public Double getAmount() {
+    public Integer getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(Integer amount) {
         this.amount = amount;
     }
-
 }
